更新

+++ guava/src/com/google/common/collect/Hashing.java
@@ -18,8 +18,7 @@ package com.google.common.collect;
 
 import com.google.common.annotations.GwtCompatible;
 import com.google.common.primitives.Ints;
-
-import javax.annotation.Nullable;
+import org.checkerframework.checker.nullness.qual.Nullable;
 
 /**
  * Static methods for implementing hash-based collections.
@@ -32,8 +31,12 @@ import javax.annotation.Nullable;
 final class Hashing {
   private Hashing() {}
 
-  private static final int C1 = 0xcc9e2d51;
-  private static final int C2 = 0x1b873593;
+  /*
+   * These should be ints, but we need to use longs to force GWT to do the multiplications with
+   * enough precision.
+   */
+  private static final long C1 = 0xcc9e2d51;
+  private static final long C2 = 0x1b873593;
 
   /*
    * This method was rewritten in Java from an intermediate step of the Murmur hash function in
@@ -44,7 +47,7 @@ final class Hashing {
    * hereby disclaims copyright to this source code.
    */
   static int smear(int hashCode) {
-    return C2 * Integer.rotateLeft(hashCode * C1, 15);
+    return (int) (C2 * Integer.rotateLeft((int) (hashCode * C1), 15));
   }
 
   static int smearedHash(@Nullable Object o) {
